+++ packages/article-image/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.1"></a>
+## [1.0.1](https://github.com/newsuk/times-components/compare/@times-components/article-image@1.0.0...@times-components/article-image@1.0.1) (2018-06-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-image
+
 <a name="1.0.0"></a>
 # [1.0.0](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.19.0...@times-components/article-image@1.0.0) (2018-06-15)