+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.3"></a>
+## [2.0.3](https://github.com/newsuk/times-components/compare/@times-components/author-profile@2.0.2...@times-components/author-profile@2.0.3) (2018-06-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-profile
+
 <a name="2.0.2"></a>
 ## [2.0.2](https://github.com/newsuk/times-components/compare/@times-components/author-profile@2.0.1...@times-components/author-profile@2.0.2) (2018-06-18)