+++ packages/image/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.1"></a>
+## [2.0.1](https://github.com/newsuk/times-components/compare/@times-components/image@2.0.0...@times-components/image@2.0.1) (2018-06-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/image
+
 <a name="2.0.0"></a>
 # [2.0.0](https://github.com/newsuk/times-components/compare/@times-components/image@1.21.0...@times-components/image@2.0.0) (2018-06-15)